XXX Got a working raw image.
This uses the same scripts then the SD card scripts found in nixpkgs, in
particular a FAT16 boot partition populated with mcopy.

I would have prefered an EXT4 partition (possibly with the /boot
directory on the same partition than /) but the tools to install a
bootloader (grub-install or extlinux) require to use losetup to mount
the disk image.

Here with a FAT16 partition, it is possible to use syslinux as
bootloader installer; it works directly on the disk image.