Add dehydrated config for noteed.com.
Dehydrated is a ACME client requesting a SSL certificate to Let's
Encrypt.

Here it is used in association to a Nginx server serving the ACME
challenge in HTTP.

Another Nginx server is used to serve HTTPS. It will fail to run until a
certificate is provided (runit will restart it until then).